# Make Kernel#then, #yield_self, #frozen? builtin