## Kernel#then, #yield_self
https://github.com/ruby/ruby/pull/3281#issuecomment-653438097

To keep Enumerator creation reasonably fast, I used builtin functions. Still it became a bit slower, but calling a block became faster, which I believe is the dominant usage of `#then` / `#yield_self`.

```
$ benchmark-driver -v --rbenv 'before;after' benchmark/kernel_then.yml --repeat-count=4
before: ruby 2.8.0dev (2020-07-03T16:52:54Z master ccb2e6eaa5) [x86_64-linux]
after: ruby 2.8.0dev (2020-07-03T18:36:08Z builtin-kernel 5fa06daca8) [x86_64-linux]
Calculating -------------------------------------
                           before       after
           kernel_then    21.343M     28.113M i/s -     20.000M times in 0.937094s 0.711421s
      kernel_then_enum    11.122M     10.756M i/s -     20.000M times in 1.798222s 1.859484s
     kernel_yield_self    21.379M     28.061M i/s -     20.000M times in 0.935492s 0.712737s
kernel_yield_self_enum    11.051M     10.781M i/s -     20.000M times in 1.809825s 1.855115s

Comparison:
                        kernel_then
                 after:  28112752.9 i/s
                before:  21342580.3 i/s - 1.32x  slower

                   kernel_then_enum
                before:  11122095.6 i/s
                 after:  10755671.9 i/s - 1.03x  slower

                  kernel_yield_self
                 after:  28060852.6 i/s
                before:  21379125.2 i/s - 1.31x  slower

             kernel_yield_self_enum
                before:  11050794.7 i/s
                 after:  10781002.7 i/s - 1.03x  slower
```

## Kernel#frozen?
```
$ benchmark-driver -v --rbenv 'before;after;before --jit;after --jit' benchmark/mjit_kernel.yml --filter=frozen --repeat-count=4
before: ruby 2.8.0dev (2020-07-03T16:52:54Z master ccb2e6eaa5) [x86_64-linux]
after: ruby 2.8.0dev (2020-07-03T18:36:08Z builtin-kernel 5fa06daca8) [x86_64-linux]
before --jit: ruby 2.8.0dev (2020-07-03T16:52:54Z master ccb2e6eaa5) +JIT [x86_64-linux]
after --jit: ruby 2.8.0dev (2020-07-03T18:36:08Z builtin-kernel 5fa06daca8) +JIT [x86_64-linux]
Calculating -------------------------------------
                         before       after  before --jit  after --jit
   mjit_frozen?(str)    39.360M     40.199M       52.619M      72.695M i/s -     40.000M times in 1.016256s 0.995045s 0.760185s 0.550247s
  mjit_frozen?(fstr)    39.881M     40.331M       52.570M      72.192M i/s -     40.000M times in 1.002995s 0.991791s 0.760887s 0.554076s

Comparison:
                mjit_frozen?(str)
         after --jit:  72694647.8 i/s
        before --jit:  52618772.3 i/s - 1.38x  slower
               after:  40199206.6 i/s - 1.81x  slower
              before:  39360167.5 i/s - 1.85x  slower

               mjit_frozen?(fstr)
         after --jit:  72192223.1 i/s
        before --jit:  52570212.1 i/s - 1.37x  slower
               after:  40331080.7 i/s - 1.79x  slower
              before:  39880547.1 i/s - 1.81x  slower
```